Firmware: add RSSI
This is a rough implementation of RSSI that reports the max, min,
average and number of RSSI measurements taken while waiting for a
DMA buffer full of symbols to be ready. This is probably not the
final implementation of RSSI, but should be enough to help create
a host-side squelch function.