le: jump to next channel earlier
Jump to the next channel in a hop pattern well in advance of when the
next packet is scheduled to arrive. This should help address issues with
clock desync between Ubertooth and
devices.

The previous behavior was to jump just moments before the next packet
was sched to arrive. If the device clock was slightly slower than
Ubertooth clock, we would jump too late and miss packets.